Add discrete contrast UI slider in the setting
This slider is temporary and will be replaced by the final UI in the next couple weeks.

In this first version, the slider is added in Settings -> Accessibility -> Color and motion. It is discrete with three values.

The slider manipulates the setting currently named CONTRAST_LEVEL in Settings.Secure. This setting is ranging from [-1, 1] with a default value of 0. However, the three values of the slider correspond to the contrast values [0, 0.5, 1] respectively

Test: Manual: put a log before Settings.System.putFloatForUser, move the slider manually and verify the logs
Test: make RunSettingsRoboTests ROBOTEST_FILTER=ContrastLevelSeekBar
Bug: 259091608
Change-Id: I2eaf751d6008ad20b3564dde2c0f4648cc6c8178

Former-commit-id: fd3911ba29a136cb6310442d825be389d3c8fd37