Make P2P Settings channel static
P2P Settings creates a P2P channel but does not close the channel if
there is a P2P connection. This causes multiple channels to accumulate
if P2P Settings is repeatedly opened and closed. Making the channel
static will ensure that each time a new P2P Settings activity is
created, the existing open channel will be used.

Bug: 259364357
Test: build
Change-Id: I999b108bc3e3c22519398a55b503078c1069cef1

Former-commit-id: b6c2ed6759fb9c124f35f854175383041c7d65a4