No screen lock notification
Settings already sends a SafetySourceIssue to Safety Center when the
user has not set a screen lock. This CL specifies the corresponding
notification for Safety Center to post for that issue.

Also specified the actionability of this issue as manual (although
this is the default so no functional change here, just for clarity).

Fix: 265799125
Bug: 260574754
Test: Manually
Change-Id: I9d579cdb8061d247e89031a5bc360f8fd5848277

Former-commit-id: ab8c9d79f44e5aa3c749e06457c860564691c70f