Remove "nosensor" restriction from FallbackHome
This was previously added in I875fa6d112db23273026dbc39b5d32748d7e99f1 to match boot animation orientation. But the first system orientation after boot is no longer influenced by FallbackHome and instead is influenced by the default rotation in DisplayRotation#readDefaultDisplayRotation so it should be safe to remove this restriction.

On the other side, if we keep this restriction it can lead to orientation flickering on devices that have ignoreOrientationRequest enabled since "nosensor" is respected on them unlike other orientations so it's possible that FallbackHome and launcher orientations won't match.

Fix: 264329911
Test: manual
Change-Id: I688a4ba3d81c1fe908b9f88c3d601aaccc4f2119

Former-commit-id: 436ede94c394832b6bd8ffe1bcb44b18dccb1c7e