Avoid disabling LE Audio when BT is off
The `updateState` function gets called when the Developer Options menua
is rendered (opened). Originally when BT is off at this point, LE Audio
is disabled, and the user needs to reboot the device again to enable LE
Audio. This CL fixes this and marks the option as disabled when BT is
off.

Bug: 263684630
Test: Manual
Change-Id: I243e2c57fc12bfefd4ab83f63336fb41639f05a0

Former-commit-id: f1007faebab8bbf4770bf182366c6e1ecb678048