Make CheckReturnValue errorprone check a warning for unit tests
CredentialManagerPreferenceControllerTest now differs between the
current development branch and internal main, and the
CheckReturnValue errorprone is failing in the branches that use
the code from the development branch. Modifying
CredentialManagerPreferenceControllerTest does not resolve this
because it results in merge conflicts, but making this check a
warning in the development branch will allow the fix to reach all
affected branches.

Fixes: 273833063
Test: m RUN_ERROR_PRONE=true SettingsUnitTests
Change-Id: Ia923f0d988b1428cbb7378f33b4703ed19113dcd

Former-commit-id: 38819fd90668149cf88fe519a503013c671dd4b7