Force MAC randomization to enabled on ACTION_WIFI_ADD_NETWORKS
The ACTION_WIFI_ADD_NETWORKS allows an app to request the user to
confirm addition of WiFi networks. The intent attaches
WifiNetworkSuggestions (used to specify the networks). However, the
calling app may configure MAC randomization to be disabled - using
reflection.

Fix to always override the MAC randomization setting to enabled. This
will override any user configuration to disable randomization - however,
considering this is (now) a new configuration it should safely start
from the default Setting.

Verified using using a test app to force non-randomized (using
reflection) value and confirm that without change randomization is
disabled and that with change it is reset correctly.

Bug: 236825742
Test: manual - use a test app
Change-Id: I3fba958c461d88005fdb5d2c3d1f2fd4ce725b8c

Former-commit-id: d3f7fa6aef24b39afd8dc332528f9c40bc68cd3d