Clarify notifications behavior in bypass setting.
This will likely be temporary as we're exploring actually displaying
notifications properly on the lockscreen (b/266423709), but that will
take a while. In the meantime, users are often confused about not seeing
notifications on the lockscreen, as that's an unexpected side effect of
enabling bypass. This should at least make that less unexpected (for
users who read the settings, at least).

I used a string already defined in a device-specific directory, and
removed that one since it applies to all devices (see other CL in
topic).

Bug: 269701672
Test: verified that the string is updated
Change-Id: I3abc6a5e1f8ee7e2f0cba3495925faf50f476e21

Former-commit-id: 8780480074f437cc1463449b9631443dbfaf5543