Changes screen flash color selected ring color in Flash Notifications
There's an outer ring to show that the screen flash color is selected and it's black in both light and dark themes.

When selecting screen flash color in the dark theme, it's not obvious which color is selected because the indicated ring is black which is similar to the dark background. Changing it to white in the dark theme for better user experience.

Bug: 275283917
Test: checks UI manually
Change-Id: I714597d99dfb74c5ce0ddef3995fd19688feacdf

Former-commit-id: 79588ff434ef10ce8d83910cee24db4cecfa4055