Fixes SelectLongPressTimeoutPreferenceControllerTest summary check.
This preference was updated in
commit 2a5b3c954769fe54a9926a9fae8df23a89f3b65c [formerly 6c3805288b15ae27d115cc89f50a672cd923870e]
to use "%s" summary and remove the controller's getSummary()
implementation so the old test that checked the controller summary
was failing.

This new version checks the preference's summary directly, but that
seems to require running using the robotest runner instead of unit test
so this test was moved there and now follows patterns used by other
similar tests there.

Fix: 278110568
Test: SelectLongPressTimeoutPreferenceControllerTest
Change-Id: Ib65c823a568ad200ebeae74ad443f32441e5cb73

Former-commit-id: 5c6ba285b70841311b5f38d91616c1e39a1f9e89